Add compatibility for React 15 RC1
This adds compatibility for React 15 RC1. The most painful/icky
change is in retrieving the internal instance, which has a random
number appended to its key.

There's also a hack to work around a bug in React 15 where warnings
on ref and key are throwing for nodes with string types.