prioritize `displayName` over `name` (#332)
React provides the ability to override a component's display name by
using the `displayName` property. When `displayName` is provided it
should take precedence over `name`. Prior to this pull request,
`displayName` would only ever be used if `name` is not present but it
should have been the other way around.