[Docs] Explicit React CSS selector syntax description
It is not obvious on reading through the document as-is that the CSS selector syntax works on React component names and their props. While it can be inferred through reading "The Key and Ref Props" section, that section can easily be missed when someone is specifically looking at the lists and examples of selectors to see what can be used. It is much clearer to explicitly state this syntax works on React components.