[enzyme-adapter-utils] [refactor] remove unused code from createMountWrapper; add tests
Technically this could be considered a breaking change since it’s a
`class` and these are public methods on it.

*However*, since nobody should be depending on this package directly,
and since enzyme itself doesn’t appear to use it, I’m going to consider
it a patch. In the event that it is a breaking change for someone, I’ll
add back the methods, ideally as no-ops.